+++ internal/xcontext/without_deadline.go
@@ -1,4 +1,4 @@
-package deadline
+package xcontext
 
 import (
 	"context"
@@ -13,7 +13,7 @@ func (valueOnlyContext) Done() <-chan struct{} { return nil }
 
 func (valueOnlyContext) Err() error { return nil }
 
-// ContextWithoutDeadline helps to clear derived deadline from deadline
-func ContextWithoutDeadline(ctx context.Context) context.Context {
+// WithoutDeadline helps to clear derived deadline from deadline
+func WithoutDeadline(ctx context.Context) context.Context {
 	return valueOnlyContext{ctx}
 }
